+++ w2vcomp/src/space/Similarity.java
@@ -3,7 +3,7 @@ package space;
 import common.MathUtils;
 
 public class Similarity {
-    public static double cosine(float[] v1, float[] v2) {
+    public static double cosine(double[] v1, double[] v2) {
         return MathUtils.cosine(v1, v2);
     }
 }
