+++ w2vcomp/src/unitTest/SigmoidTest.java
@@ -14,13 +14,13 @@ public class SigmoidTest {
     @Test
     public void testPrecision() {
         Random rand = new Random();
-        float maxX = 10;
+        double maxX = 10;
         SigmoidTable sigmoidComputer = new SigmoidTable(1000000, maxX);
-        float delta = (float) Math.exp(-12);
+        double delta = (double) Math.exp(-12);
         for (int i = 0; i < 10000; i++) {
-            float x = rand.nextFloat() * 2 * maxX - maxX;
-            float realSigmoid = (float) MathUtils.sigmoid(x);
-            float precomputedSigmoid = sigmoidComputer.getSigmoid(x);
+            double x = rand.nextFloat() * 2 * maxX - maxX;
+            double realSigmoid = (double) MathUtils.sigmoid(x);
+            double precomputedSigmoid = sigmoidComputer.getSigmoid(x);
             assertEquals(precomputedSigmoid, realSigmoid, delta);
         }
     }
