+++ w2vcomp/src/word2vec/SkipNGramWord2Vec.java
@@ -10,15 +10,15 @@ import vocab.VocabEntry;
  */
 public class SkipNGramWord2Vec extends SingleThreadWord2Vec {
     public SkipNGramWord2Vec(int projectionLayerSize, int windowSize,
-            boolean hierarchicalSoftmax, int negativeSamples, float subSample) {
+            boolean hierarchicalSoftmax, int negativeSamples, double subSample) {
         super(projectionLayerSize, windowSize, hierarchicalSoftmax,
                 negativeSamples, subSample);
     }
 
     public void trainSentence(int[] sentence) {
         // train with the sentence
-        float[] a1 = new float[projectionLayerSize];
-        float[] a1error = new float[projectionLayerSize];
+        double[] a1 = new double[projectionLayerSize];
+        double[] a1error = new double[projectionLayerSize];
         int sentenceLength = sentence.length;
         int iWordIndex = 0;
         for (int wordPosition = 0; wordPosition < sentence.length; wordPosition++) {
@@ -53,7 +53,7 @@ public class SkipNGramWord2Vec extends SingleThreadWord2Vec {
                     // HIERARCHICAL SOFTMAX
                     if (hierarchicalSoftmax) {
                         for (int bit = 0; bit < word.code.length(); bit++) {
-                            float z2 = 0;
+                            double z2 = 0;
                             int iParentIndex = word.ancestors[bit];
                             // Propagate hidden -> output
                             for (int j = 0; j < projectionLayerSize; j++) {
@@ -61,12 +61,12 @@ public class SkipNGramWord2Vec extends SingleThreadWord2Vec {
                                         * weights1[iParentIndex][j];
                             }
 
-                            float a2 = sigmoidTable.getSigmoid(z2);
+                            double a2 = sigmoidTable.getSigmoid(z2);
                             if (a2 == 0 || a2 == 1)
                                 continue;
                             // 'g' is the gradient multiplied by the learning
                             // rate
-                            float gradient = (float) ((1 - (word.code
+                            double gradient = (double) ((1 - (word.code
                                     .charAt(bit) - 48) - a2) * alpha);
                             // Propagate errors output -> hidden
                             for (int j = 0; j < projectionLayerSize; j++) {
@@ -100,14 +100,14 @@ public class SkipNGramWord2Vec extends SingleThreadWord2Vec {
                                     continue;
                                 label = 0;
                             }
-                            float z2 = 0;
-                            float gradient;
+                            double z2 = 0;
+                            double gradient;
                             for (int j = 0; j < projectionLayerSize; j++) {
                                 z2 += weights0[iWordIndex][j]
                                         * negativeWeights1[target][j];
                             }
-                            float a2 = sigmoidTable.getSigmoid(z2);
-                            gradient = (float) ((label - a2) * alpha);
+                            double a2 = sigmoidTable.getSigmoid(z2);
+                            gradient = (double) ((label - a2) * alpha);
                             for (int j = 0; j < projectionLayerSize; j++) {
                                 a1error[j] += gradient
                                         * negativeWeights1[target][j];
